centralise Pool implementation as well
