+++ baselines/nqueens/nqueens_multigpu_cuda.cu
@@ -12,73 +12,11 @@
 #include "cuda_runtime.h"
 
 #include "lib/NQueens_node.h"
+#include "lib/Pool.h"
 
 #define BLOCK_SIZE 512
 
 #define MIN(a, b) ((a) < (b) ? (a) : (b))
-
-/*******************************************************************************
-Implementation of a dynamic-sized single pool data structure.
-Its initial capacity is 1024, and we reallocate a new container with double
-the capacity when it is full. Since we perform only DFS, it only supports
-'pushBack' and 'popBack' operations.
-*******************************************************************************/
-
-#define CAPACITY 1024
-
-typedef struct
-{
-  Node* elements;
-  int capacity;
-  int front;
-  int size;
-} SinglePool;
-
-void initSinglePool(SinglePool* pool)
-{
-  pool->elements = (Node*)malloc(CAPACITY * sizeof(Node));
-  pool->capacity = CAPACITY;
-  pool->front = 0;
-  pool->size = 0;
-}
-
-void pushBack(SinglePool* pool, Node node)
-{
-  if (pool->front + pool->size >= pool->capacity) {
-    pool->capacity *= 2;
-    pool->elements = (Node*)realloc(pool->elements, pool->capacity * sizeof(Node));
-  }
-
-  pool->elements[pool->front + pool->size] = node;
-  pool->size += 1;
-}
-
-Node popBack(SinglePool* pool, int* hasWork)
-{
-  if (pool->size > 0) {
-    *hasWork = 1;
-    pool->size -= 1;
-    return pool->elements[pool->front + pool->size];
-  }
-
-  return (Node){0};
-}
-
-Node popFront(SinglePool* pool, int* hasWork)
-{
-  if (pool->size > 0) {
-    *hasWork = 1;
-    pool->size--;
-    return pool->elements[pool->front++];
-  }
-
-  return (Node){0};
-}
-
-void deleteSinglePool(SinglePool* pool)
-{
-  free(pool->elements);
-}
 
 /*******************************************************************************
 Implementation of the multi-GPU N-Queens search.
